% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Cars Dataset}
\format{
A data frame with 50 rows and 2 variables:
\describe{
\item{speed}{A numeric value representing how fast the car was going (in mph).}
\item{dist}{A numeric value for how far the car traveled before stopping (in feet).}
}
}
\source{
Built-in R dataset.
}
\usage{
cars
}
\description{
This dataset contains the speed of cars and the corresponding stopping distances.
}
\examples{
data(cars)
head(cars)
}
\keyword{datasets}
