% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mymean}
\alias{mymean}
\title{Compute the Mean of a Numeric Vector}
\usage{
mymean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean of the input numeric vector.
}
\description{
This function calculates the mean of a given numeric vector.
}
\examples{
mymean(c(12,20,18,11))
}
